Fix: QuadrangleGeometricPiece: computation of specular reflection
The strength of specular reflection should be calculated from
the Z component (= component for the direction of the user's gaze)
of the vector of the reflected light.

The former implementation had calculated the strength simply from
the inner product of the normal vector and the vector of the light source,
ignoring the optical mechanism of the specular reflection,
so it looked unnatural a little.

Hence, by this commit,
we have fixed the calculation of the specular reflection.