Add: SimpleRenderer: CAS operations for screen-updated/resized flags
Following two flags have been added:

* screenUpdated:
    represents whether the content of the screen has been updated.

* screenResized:
    represents whether the screen has been resized.

An app-side thread (to be implemented) refers these flags periodically.
For example, if the value of "screenUpdated" flag is true,
the app-side thread updates the screen, and then resets the flag to false.

However, user's code running on an other thread may call render() method,
and the updating of the flag caused by it may conflict to the above.

Hence, operations for referencing and changing the value of this flag
must be performed atomically.
The same for "screenResized" flag.

So we have implemented Compare-and-Swap (CAS) operation methods for them,
instead of setter/getter methods.