Remove: MainWindow: "synchronized" modifier from some methods
We have removed the "synchronized" modifier from UI-operation methods,
such as resize(), repaintScreen(), setScreenImage(image), etc.

The internal processing of the UI operation methods are
always processed in serial, on the event-dispatcher thread.
So "synchronized" is not necessary for them.

In addition, by the existence of the above "synchronized",
we have been required to be careful not to call
other synchronized methods of the same instance,
from other tasks stacked on the event dispatcher's queue
(Otherwise the thread may fall into a deadlock).
It was the waste of time for thinking about the above situation
because the "synchronized" is not necessary as mentioned above.

Hence, we have removed it by this commit.