Remove: RinearnGraph3D: unnecessary "sleep"
We inserted "sleep" just after popping-up
an error message window in the constructor,
just to make sure to popping-up it.

Because sometimes it didn't pop up,
and we thought that IllegalStateException is thrown
before popping up the window,
caused by some kind of time-lag.

However, probably, it was a wrong guess.
Probably we took a mistake for condition-expression of
"if" statement, when we implemented and debugged this part.

Therefore we removed the unnecessary "sleep".