Refactor: Model: make some fields "public final"
Following fields have been changed to \"public final\" from \"private\":

* config
* scriptEngineMount

Because:

Changing the references of the above fields might be cause of confusing,
when the internal structure of this app becomes more complex.

In addition, considering any situation, probably,
it is not necessary to change them.

Therefore we want to make their references unmodifiable.