Add: API: getImage()
Gets an Image instance storing the current screen image.

This method copies the current screen image to a buffer,
and returns the reference to the buffer.
Hence, the content of the returned Image instance is
NOT updated automatically when the screen is re-rendered.
To update the content of the image, re-call this method.